+++ Examples/FKKitExamples/FKKitExamples/Examples/FKUIKit/TabBar/Hub/FKTabBarExamplesHubViewController.swift
@@ -275,6 +275,10 @@ final class FKTabBarExamplesHubViewController: UITableViewController {
     ),
   ]
 
+  convenience init() {
+    self.init(style: .insetGrouped)
+  }
+
   override func viewDidLoad() {
     super.viewDidLoad()
     title = "TabBar"